+++ src/com/sun/labs/minion/indexer/postings/io/StreamPostingsOutput.java
@@ -77,10 +77,10 @@ public class StreamPostingsOutput implements PostingsOutput {
      * @return The number of bytes written.
      */
     public int write(WriteableBuffer b) throws java.io.IOException {
-        int size = (int) b.position();
+        long size = b.position();
         b.write(stream);
         offset += size;
-        return size;
+        return (int) size;
     }
     
     /**
